pyproject.toml
[tool.poetry]
name = "paper-bot"
version = "0.1.0"
description = ""
authors = ["Nikita Yatchenko <yatchenko.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.10"
unstructured = {extras = ["all-docs"], version = "0.17.2"}
pillow = "^11.2.1"
lxml = "^5.3.2"
matplotlib = "^3.10.1"
tiktoken = "^0.9.0"
langchain = "^0.3.23"
openai = "^1.75.0"
langchain-openai = "^0.3.14"
langchain-chroma = "^0.2.3"
langchain-experimental = "^0.3.4"
aiogram = "^3.20.0.post0"

[tool.poetry.group.dev.dependencies]
jupyter = "1.1.1"
isort = "6.0.1"
flake8 = "7.1.2"
pre-commit = "4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
